chore: put validators into separate folder
