+++ packages/backend/middlewares/index.js
@@ -1,8 +1,8 @@
 /* eslint global-require: 0 */
 module.exports = {
   ...require('./auth'),
-  ...require('./studentValidator'),
-  ...require('./teacherValidator'),
-  ...require('./messageValidator'),
-  ...require('./directMessageValidator'),
+  ...require('./validators/studentValidator'),
+  ...require('./validators/teacherValidator'),
+  ...require('./validators/messageValidator'),
+  ...require('./validators/directMessageValidator'),
 };
